Change note in note container to array of notes, handle removing note from array
This code uses Output and EventEmitter to handle removing a note from
the notes array. When the check is clicked, that event gets passed
up from the note-card to the note container and handles removing it from
the array of notes using the index of the note in the array.